Extract classes into own files
