+++ YouseiReloaded/Serialization/Yaml/YamlConfig.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+using Yousei.Shared;
+
+namespace YouseiReloaded.Serialization.Yaml
+{
+    internal class YamlConfig
+    {
+        public Dictionary<string, Dictionary<string, object>> Connections { get; init; } = new();
+
+        public Dictionary<string, FlowConfig> Flows { get; init; } = new();
+    }
+}
